refactor: rename AccountRestrictionNotice to TemporaryBanAlert
Update references and move component files.

#VERCEL_SKIP
