--- components/account-restriction-notice/index.ts
@@ -1,1 +1,0 @@
-export { AccountRestrictionNotice } from "./account-restriction-notice"
