+++ components/account-restriction-notice/temporary-ban-alert.tsx
