+++ components/temporary-ban-alert/index.ts
@@ -1,0 +1,1 @@
+export { TemporaryBanAlert } from "./temporary-ban-alert"
